Increased OF 1.3 flow table size to 512

+++ ZodiacFX/src/openflow/openflow.h
@@ -37,10 +37,10 @@
 
 struct flows_counter
 {
-	int hitCount;
+	uint16_t hitCount;
 	int bytes;
 	int duration;
-	int active;
+	uint8_t active;
 	int lastmatch;
 };
